chore: make check baştan sona yeşil — sıfır uyarı, biçim, macOS IWYU
Sıfır uyarı: painter_backend, symbol_preview, theme ve settings_dialog'da
on adet -Wdouble-promotion / -Wunused-lambda-capture kapatıldı. Anayasa
6.3 uyarısız bir yapı istiyor; kimsenin okumadığı uyarılarla dolu bir
yapı içindeki gerçek bulguyu da gizler.

Biçim: yeniden adlandırma tanımlayıcıları kısalttığı için clang-format'ın
sarma noktaları kaydı ve ağaç 519 ihlale çıkmıştı, make format ile
sıfırlandı. Daha önce "yeniden biçimlendirmek CI'ı kırar" demiştim; bu
yanlıştı — CI'da clang-format işi hiç yok, yalnızca ci-gate-*.sh
çalışıyor. Sürüm hâlâ hiçbir yerde sabitlenmiş değil ve bu ayrı bir iş.

IWYU macOS'ta: iwyu_tool.py SDK yolunu bilmediği için ilk dosyada
'type_traits' file not found ile düşüyordu — bozuk bir checkout gibi
okunan bir toolchain iletisi. scripts/run-iwyu.sh xcrun --show-sdk-path
ekliyor; Article 10 gereği Makefile'da OS koşulu değil, betikte.

make check rc=0: 22 gate, 517 vaka, 7736 doğrulama, sıfır uyarı.

+++ src/app/src/painter_backend.cpp
@@ -394,7 +394,7 @@ private:
         // under the forest triangles. A pattern layer paints its pattern.
 
         QPen pen(faded(ps.line_rgba, ps.opacity));
-        pen.setWidthF(ps.line_width_px);
+        pen.setWidthF(static_cast<qreal>(ps.line_width_px));
         painter.setPen(pen);
 
         // Rotated about the box centre so the pattern is continuous across the
@@ -439,7 +439,7 @@ private:
         // a `dolgu` layer underneath the pattern, which is what draws the green
         // under the forest triangles. A pattern layer paints its pattern.
 
-        painter.setPen(QPen(faded(ps.line_rgba, ps.opacity), ps.line_width_px));
+        painter.setPen(QPen(faded(ps.line_rgba, ps.opacity), static_cast<qreal>(ps.line_width_px)));
         painter.setBrush(glyph_brush(ps));
 
         // Anchored to the world grid rather than to the bounding box, so the
@@ -468,7 +468,7 @@ private:
         if (batch.runs.empty()) return;
 
         const double size = ps.size_px > 0.5f ? static_cast<double>(ps.size_px) : 6.0;
-        painter.setPen(QPen(faded(ps.line_rgba, ps.opacity), ps.line_width_px));
+        painter.setPen(QPen(faded(ps.line_rgba, ps.opacity), static_cast<qreal>(ps.line_width_px)));
         painter.setBrush(glyph_brush(ps));
 
         std::size_t offset = 0;
@@ -787,7 +787,7 @@ private:
         if (batch.runs.empty()) return;
 
         QPen pen(faded(batch.rgba, ps.opacity));
-        pen.setWidthF(batch.width_px);
+        pen.setWidthF(static_cast<qreal>(batch.width_px));
         pen.setCapStyle(qt_cap(ps.cap));
         pen.setJoinStyle(qt_join(ps.join));
         apply_dash(pen, ps);
@@ -818,7 +818,7 @@ private:
         const double size = ps.size_px > 0.5f ? static_cast<double>(ps.size_px) : 5.0;
         const double interval =
             ps.interval_px > 0.5f ? static_cast<double>(ps.interval_px) : size * 3.0;
-        painter.setPen(QPen(faded(ps.line_rgba, ps.opacity), ps.line_width_px));
+        painter.setPen(QPen(faded(ps.line_rgba, ps.opacity), static_cast<qreal>(ps.line_width_px)));
         // A hash tick is a stroke and has no interior to fill.
         painter.setBrush(hash ? QBrush(Qt::NoBrush) : glyph_brush(ps));
 
@@ -1031,7 +1031,7 @@ private:
             if (batch.runs.empty()) continue;
 
             QPen pen(from_rgba(batch.rgba));
-            pen.setWidthF(batch.width_px);
+            pen.setWidthF(static_cast<qreal>(batch.width_px));
             pen.setStyle(batch.dashed ? Qt::DashLine : Qt::SolidLine);
             painter.setPen(pen);
             painter.setBrush(batch.fill_rgba != 0 ? QBrush(from_rgba(batch.fill_rgba))
